requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.13
PyNaCl>=1.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
